WinGui: Fix a second Integer overflow.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@4917 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ win/CS/HandBrakeWPF/ViewModels/PictureSettingsViewModel.cs
@@ -1004,7 +1004,7 @@ namespace HandBrakeWPF.ViewModels
 
                         double new_width = ((double)this.Height * this.sourceResolution.Height * this.SourceAspect.Width *
                                             crop_width) /
-                                           (this.sourceResolution.Width * this.SourceAspect.Height * crop_height);
+                                           ((double)this.sourceResolution.Width * this.SourceAspect.Height * crop_height);
 
                         this.Task.Width = (int)Math.Round(this.GetModulusValue(new_width), 0);
                         this.NotifyOfPropertyChange(() => this.Task.Width);
